+++ CHANGES.md
@@ -10,6 +10,7 @@
  * removed extra line endings from `remote_version`
  * added ratings function for creating star ratings based upon GitHub repo data.
  * bring parts of `class GitHub_Theme_Updater` code on par with `class GitHub_Plugin_Updater`
+ * added 'ghu-' prefix to transients
 
 #### 2.1.1